Fix memory leak when setting variables
I wasn't sure if it was safe to do this but after reading the tesseract source code I think we can reasonably assume that tesseract does not require the string used to set a variable to stay around forever. I'm certain that almost all code (including bindings to other languages) relies on this undocumented implementation detail for safety. Let's hope they don't change it too soon.